FIX: taking out all console.log. FEATURE: !ID for id routes.

+++ src/services/deleteProductService.ts
@@ -1,10 +1,11 @@
 import { database } from "../database";
 
-export const deleteProductService = async (id : string) => {
-    try {
-        const deletedProduct = await database.findOneAndDelete({_id: id});
-        return deletedProduct;
-    } catch (error: any) {
-        throw new Error(error.message);
-    }
+export const deleteProductService = async (id: string) => {
+  try {
+    if (!id) throw new Error("Missing ID parameter");
+    await database.findOneAndDelete({ _id: id });
+    return { message: "Product deleted successfully" };
+  } catch (error: any) {
+    throw new Error(error.message);
+  }
 };
